Conflicts:
	webroot/subpages/documents/forview.php

+++ src/Template/Clients/add.ctp
@@ -50,7 +50,7 @@
             <a href="javascript:window.print();" class="floatright btn btn-info">Print</a>
         <?php } 
        
-        if (isset($client) && $sidebar->client_delete == '1') { ?>
+        if (isset($client) && $sidebar->client_delete == '1' && $param != 'add') { ?>
             <a href="<?php echo $this->request->webroot; ?>clients/delete/<?php echo $client->id; ?><?php echo (isset($_GET['draft']))?"?draft":""; ?>"
                onclick="return confirm('Are you sure you want to delete <?= h($client->company_name) ?>?');"
                class="floatright btn btn-danger btnspc">Delete</a>
@@ -99,77 +99,6 @@
                     </div>
                 </div>
 
-                <?php
-                    /*if ($this->request->params['action'] == 'edit')
-                        include('subpages/clients/recruiter_contact_table.php');*/
-                    if (isset($_GET['view'])) {
-                        ?>
-                        <table class="table table-striped table-bordered table-advance table-hover recruiters">
-                            <thead>
-                            <tr>
-                                <th colspan="2">Assigned To:</th>
-                            </tr>
-                            </thead>
-                            <tbody id="">
-                            <?php
-                                $types = array('Driver', 'Admin', 'Recruiter', 'External', 'Safety', 'Driver', 'Contact', 'Employee', 'Guest', 'Partner');
-                                $counter = 0;
-                                foreach ($getprofile as $p) {
-                                    ?>
-                                    <tr>
-                                        <td>
-                                            <?php
-                                                echo '<a href="' . $this->request->webroot . 'profiles/view/' . $p->id . '">';
-                                                if (strlen($p->username)>0) {
-                                                    echo $p->username;
-                                                } elseif(strlen($p->fname)>0 or strlen($p->lname)>0) {
-                                                    echo $p->fname . " " . $p->lname;
-                                                } else {
-                                                    echo "[NO NAME]";
-                                                }
-                                                $profiletype = " (Draft)";
-                                                if (strlen($p->profile_type)>0 ) {
-                                                    if ($p->profile_type < count($types)) {
-                                                        $profiletype = " (" . $types[$p->profile_type] . ")";
-                                                    } else {
-                                                        $profiletype = " (UNKNOWN)";
-                                                    }
-                                                }
-                                                echo $profiletype;
-
-                                                echo "</a>"
-                                                //<a href="<?php echo $this->request->webroot;!>profiles/view/<?php echo $p->id; !>">
-                                            //<?php echo $p->username; !> (<?php echo $types[$p->profile_type]; !>)&nbsp;&nbsp;(Profile)</a>
-
-                                            ?>
-                                        </td>
-                                    </tr>
-                                    <?php
-                                    $counter++;
-                                }
-                                $c = $counter;
-                                if ($counter == 0) {
-                                    echo "<TR><TD>No <?php echo strtolower($settings->profile); ?>s</TD></TR>";
-                                }
-                            ?>
-                            <?php
-                                foreach ($getcontact as $cont) {
-                                    ?>
-                                             <tr><td>
-												<a href="<?php echo $this->request->webroot;?>profiles/view/<?php echo $cont->id; ?>">
-												    <?php echo $cont->username; ?> <?php //echo $types[$p->profile_type];
-                                    ?>&nbsp;&nbsp;(Contact)
-                                                </a>
-											</tr></td>
-                                                    <?php
-                                }
-                            ?>
-                            </tbody>
-                        </table>
-                    <?php
-                    }
-                ?>
-
             </div>
             <div class="col-md-9">
 
@@ -191,13 +120,13 @@
                                 <li>
                                     <a href="#tab_1_2" data-toggle="tab">Display</a>
                                 </li>
-                                 <li>
-                                    <a href="#tab_1_3" data-toggle="tab">Assign to Profile</a>
-                                </li>
+                                 
                             <?php
 
                             } ?>
-
+                                <li>
+                                    <a href="#tab_1_3" data-toggle="tab"><?php echo (!isset($_GET['view']))?"Assign to Profile":"Assigned To";?></a>
+                                </li>
 
                         </UL>
                     </div>
@@ -628,12 +557,13 @@
 
                             </form>
                         </div>
+                        <?php } ?>
                          <div class="tab-pane" id="tab_1_3">
                          <?php
                             include('subpages/clients/recruiter_contact_table.php');
                          ?>
                          </div>
-                        <?php } ?>
+                        
 
                         <!-- END SAMPLE FORM PORTLET-->
 </div>
